# conveyor/src/main/java/com/github/nst1610/neoflex/project/conveyor/model/PaymentScheduleElement.java
package com.github.nst1610.neoflex.project.conveyor.model;

import lombok.AllArgsConstructor;
import lombok.Data;

import java.math.BigDecimal;
import java.time.LocalDate;

@Data
@AllArgsConstructor
public class PaymentScheduleElement {
    private Integer number;
    private LocalDate date;
    private BigDecimal totalPayment;
    private BigDecimal interestPayment;
    private BigDecimal debtPayment;
    private BigDecimal remainingDebt;
}